More PR Cleanup

+++ Code/Views/LYRUIMessageBubbleView.h
@@ -27,7 +27,7 @@ extern CGFloat const LYRUIMessageBubbleDefaultHeight;
 /**
  @abstract Tells the bubble view to display a download indicator on top of content.
  */
-- (void)updateActivityIndicatorWithProgress:(float)progress style:(LYRUIProgressViewIconStyle)style;
+- (void)updateActivityIndicatorWithProgress:(double)progress style:(LYRUIProgressViewIconStyle)style;
 
 /**
  @abstract Tells the bubble view to display a given string.
